fix bugs: .pth file can't find the models.py, fix it with copy models.py to the dir

+++ server/matting_model.py
@@ -242,7 +242,7 @@ def model_infer_video(model, app_info, data):
             com = fgr * pha + tgt_bgr * (1-pha)
             
             if (opt % 2 == 1):
-                com = STF_model(com * 255)
+                com = STF_model(com.to(torch.float16) * 255)
                 com =  (com.clamp(-1,1) + 1) * 0.5
